+++ Compositor/lib/Mesa/include/IOutput.h
@@ -36,7 +36,7 @@ namespace Compositor {
              * @param sequence Commit sequence number that was presented.
              * @param time Presentation time stamp, 0 means never presented/timeout.
              */
-            virtual void Presented(const IOutput* output, const uint32_t sequence, const uint64_t time) = 0;
+            virtual void Presented(const IOutput* output, const uint64_t sequence, const uint64_t time) = 0;
         }; // struct ICallback
 
         /*
